+++ Tests/Monads.Tests.Unit/Usings.cs
@@ -1,1 +1,4 @@
+global using FluentAssertions;
 global using Xunit;
+global using static Monads.Tests.Unit.MaybeTests.Helpers.MaybeHelperMethods;
+global using Frognar.Monads;
